organize: use a new file for each command
